added test for mean_data_stack
